- Issue #75: Backup and/or Copy Videos
Enable from the Main Menu a function to Backup or Copy the Videos folder to a
designated location (e.g. internal storage, SD card, OTG flash drive).

I added a new menu item to the main menu for copying of all videos.  The
VideoActivity is very similar to the ShareActivity and tehy share a lot of the
code, especially the adapter class for managing the list of videos.

In an attempt to share code, I moved most of the functionality from ShareActivity
to a new class, VideoListHelper.  ExportedVideosAdapter is shared between both
Share and Copy activites.

UI Changes for the Share Phase:
   - The phase has been renamed to "View + Share"
   - The video folders button has been removed, since the VideoActivity provides
     the functionality to copy videos to a new location, especially an external drive.

UI for the Copy/Share Files activity:
  - The list of video files is now capable for multi-selection.  Each of the
    videos is displayed as a checkbox.
  - The share button exists for each video so that it can be shared to other
    applications from this activity.  There is no delete button on the videos.
  - When files are selected, the copy button is enabled so the user
    can copy multiple files at the same time.
  - The labels for 3gp and mp4 files at thhave been changed to checkboxes to allow
    easy selection of all files of type 3gp and mp4
  - When you click the copy button, it checks that you had previously selected
    a copy folder and that the folder exists (since it could be on a removable
    drive)  If there is no copy folder available, then it goes directly to the
    folder specification UI.  If the copy folder was available, then it puts
    up a message that allows the user to copy using the existing folder or to
    specify a new folder.  If the video file already exists in the copy directory,
    warn the user before overwriting it.

Other Changes-
- Split out methods from MainActivity to MainBaseActivity that can be shared
across other activities for a more uniform UI across activity pages.  Initially
the new class is used as the basis for VideoActivity.
MainBaseActivity shares the UI for the main menu, the drawer, and help screen

- ExportedVideosAdapter was creating a new view for each item every time getView
was called.  Since this is called multiple times during geometry management for
the page, it meant that there were lots of copies of the views for each item.
This becomes important for the checkbox UI, because new views were getting
swapped in and checked checkboxes were suddenly becoming unchecked.

- Workspace will persist the copyPath in the same way that it persists the
workspace path.  This allows the copyPath to be preserved between sessions
with the intent that the user will be copying videos to the same folder
usually and it will minimize their visits to the find folder directory UI.

- UriUtils code has been improved to better differentiate uri's that are
referencing internal memory vs sdcard memory.  I attempted to include
usb drives as well, but that code is still a work in progress.  It is good
enough for current purposes, but needs more improvement.

- New Unit Tests TestSharePhase and TestVideoActivity