- Issue #75: Backup, Delete, Resore Story Projects - Part 2
Enable from the Main Menu a function to Backup/Restore/Delete story folders
(aka projects) to a designated location (e.g. internal storage, SD card, OTG
flash drive).

I added a new menu item to the main menu for doing backup of story projects.
Visually, the BackupRestorectivity is very similar to MainActivity (aka Story
Templates) from which it is extended.  Therefore they share a lot of the code.

Of the new classes BackupRestoreActivity extends from MainActivity,
BackupRestoreAdapter extends StoryPageAdapter, and BackupStoryPageFragment
extends StoryPageFragment

Visually, in BackupRestoreActivity the list of stories is the same as MainActivty,
except that the picture icon at the left has been replaced with a checkbox to
indicate that the items are selectable.  There are buttons that allow the
selected stories to be either delted or backed up.  The backup will default to
the user selected backup directory (or he can select a new backup directory).
Deletion of a story will need to be confirmed.  For backup, if a copy of the
story already exists the user will be warned before the original copy is removed.
To restore a story project, the user will be asked to identify the location of
the story to be restored.